refactor: 예외 클래스 효율적으로 관리하기 위해 폴더 구조 변경
