+++ src/main/java/softeer2nd/chess/exception/moveException/InvalidDistanceException.java
@@ -1,4 +1,4 @@
-package softeer2nd.chess.exception;
+package softeer2nd.chess.exception.moveException;
 
 public class InvalidDistanceException extends InvalidMoveException {
     public InvalidDistanceException(String message) {
