+++ src/main/java/softeer2nd/chess/exception/moveException/InvalidMoveException.java
@@ -1,4 +1,4 @@
-package softeer2nd.chess.exception;
+package softeer2nd.chess.exception.moveException;
 
 public class InvalidMoveException extends RuntimeException {
     public InvalidMoveException(String message) {
