+++ src/main/java/softeer2nd/chess/exception/pieceException/InvalidColorException.java
@@ -1,4 +1,4 @@
-package softeer2nd.chess.exception;
+package softeer2nd.chess.exception.pieceException;
 
 public class InvalidColorException extends InvalidPieceFieldException {
     public InvalidColorException() {
