+++ src/main/java/softeer2nd/chess/exception/pieceException/InvalidPieceFieldException.java
@@ -1,4 +1,4 @@
-package softeer2nd.chess.exception;
+package softeer2nd.chess.exception.pieceException;
 
 public class InvalidPieceFieldException extends IllegalArgumentException {
     public InvalidPieceFieldException(String message) {
