+++ src/main/java/softeer2nd/chess/exception/pieceException/InvalidTypeException.java
@@ -1,4 +1,4 @@
-package softeer2nd.chess.exception;
+package softeer2nd.chess.exception.pieceException;
 
 public class InvalidTypeException extends InvalidPieceFieldException {
     public InvalidTypeException() {
