+++ src/main/java/softeer2nd/chess/exception/positionException/InvalidBoardBoundException.java
@@ -1,4 +1,4 @@
-package softeer2nd.chess.exception;
+package softeer2nd.chess.exception.positionException;
 
 public class InvalidBoardBoundException extends RuntimeException{
     public InvalidBoardBoundException(String message) {
