+++ src/main/java/softeer2nd/chess/exception/positionException/InvalidPositionException.java
@@ -1,4 +1,4 @@
-package softeer2nd.chess.exception;
+package softeer2nd.chess.exception.positionException;
 
 public class InvalidPositionException extends RuntimeException {
     public InvalidPositionException(String message) {
