+++ src/test/java/softeer2nd/chess/board/PositionTest.java
@@ -3,7 +3,7 @@ package softeer2nd.chess.board;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import softeer2nd.chess.Board.Position;
-import softeer2nd.chess.exception.InvalidPositionException;
+import softeer2nd.chess.exception.positionException.InvalidPositionException;
 
 import static org.junit.jupiter.api.Assertions.*;
