+++ src/test/java/softeer2nd/chess/pieces/PieceFactoryTest.java
@@ -2,7 +2,7 @@ package softeer2nd.chess.pieces;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import softeer2nd.chess.exception.InvalidPieceFieldException;
+import softeer2nd.chess.exception.pieceException.InvalidPieceFieldException;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static softeer2nd.chess.Board.Constants.*;
